[WebDriver BiDi] fix `test_release_mouse_sequence_resets_dblclick_state`
